resume screener/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
python-dotenv
